move accum_row_inplace to diskann-utils; demote get_chunk_from_training_data to test helper
