+++ diskann-providers/src/model/pq/mod.rs
@@ -10,10 +10,9 @@ pub use fixed_chunk_pq_table::{
 
 mod pq_construction;
 pub use pq_construction::{
-    MAX_PQ_TRAINING_SET_SIZE, NUM_KMEANS_REPS_PQ, NUM_PQ_CENTROIDS, accum_row_inplace,
-    generate_pq_data_from_pivots, generate_pq_data_from_pivots_from_membuf,
-    generate_pq_data_from_pivots_from_membuf_batch, generate_pq_pivots,
-    generate_pq_pivots_from_membuf, get_chunk_from_training_data, move_train_data_by_centroid,
+    MAX_PQ_TRAINING_SET_SIZE, NUM_KMEANS_REPS_PQ, NUM_PQ_CENTROIDS, generate_pq_data_from_pivots,
+    generate_pq_data_from_pivots_from_membuf, generate_pq_data_from_pivots_from_membuf_batch,
+    generate_pq_pivots, generate_pq_pivots_from_membuf, move_train_data_by_centroid,
 };
 
 /// all metadata of individual sub-component files is written in first 4KB for unified files
